코드 컨벤션에 맞게 디렉토리 구조 변경 및 services 코드 생성
