+++ package.json
@@ -1,10 +1,10 @@
 {
   "name": "1-howdoilook-1team",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "app.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "nodemon curating/app.js",
+    "dev": "nodemon src/app.js",
     "seed": "prisma db seed"
   },
   "prisma": {
