+++ src/routes/curationRoutes.js
